Add the IS_DEVELOPMENT_PREVIEW environment variable. (#85)
This PR adds the environment variable to a few of the CI scripts.
The variable toggles 'development' or 'production' mode in the
Theme's Webpack toolchain. By setting the variable to false, we
ensure the toolchain runs in 'production' mode. If set to true, the
toolchain is run in 'development' mode.

J=SLAP-962
TEST=manual

Ran the modified CI scripts and ensured the assets were generated
in the correct mode in each case.