Add a patch file that provides an .npmrc (#103)
This patch adds an .npmrc to the repository. This file configures how npm is run in the CBD container. We set `engine-strict=true` so that npm will compare the container's Node version to the minimum Node version required by Jambo.

TEST=manual

Cloned an older version of the template repo that did not have the .npmrc. I ran the `ci/install_deps.sh` script to install all of the dependencies. I then applied this patch, making sure I was using Node 10. I attempted `npm upgrade jambo` but was given a helpful error informing me that my Node version was incompatible with the one required by Jambo v1.10.4. After changing to Node v12, the upgrade worked as expected.