Bump node version to 16 (#185)
Bump node version to 16

The `base_image_tag` is deprecated, and the current way of setting the node version is with the `.node-version` file. 

Note: This will cause the package-lock to update to version 2. We will wait until the next Theme release (v1.29) to merge this in.

J=SLAP-2048
TEST=manual

Test this with the latest theme version (v1.28) and confirm in the console that node 16 is being used. Add a universal page and see the page served correctly with live preview